Bug 1926357: Update wasm spec tests to wasm-3.0. r=rhunt,supply-chain-reviewers
Many proposals have been merged into the wasm-3.0 branch of the
WebAssembly spec repo. It currently represents the state of the art of
wasm, which we keep up with - proposals that are officially standardized
but not yet super widely implemented.

Basing our spec tests on this reduces duplication across proposal repos
and should help us catch spec problems earlier.

This patch reworks the spec test generator slightly to handle some new
module definition syntax, and removes proposals which have already been
merged into wasm-3.0.

Differential Revision: https://phabricator.services.mozilla.com/D227573

+++ js/src/jit-test/tests/wasm/spec/spec/unreachable.wast.js
@@ -22,8 +22,8 @@ let $0 = instantiate(`(module
   (func $$dummy3 (param i32 i32 i32))
 
   (func (export "type-i32") (result i32) (unreachable))
-  (func (export "type-i64") (result i32) (unreachable))
-  (func (export "type-f32") (result f64) (unreachable))
+  (func (export "type-i64") (result i64) (unreachable))
+  (func (export "type-f32") (result f32) (unreachable))
   (func (export "type-f64") (result f64) (unreachable))
 
   (func (export "as-func-first") (result i32)
